more sol8 fixes

+++ smart-contracts/contracts/UnlockUtils.sol
@@ -49,7 +49,7 @@ library UnlockUtils {
   ) internal pure
     returns(string memory)
   {
-    bytes32 value = bytes32(uint256(_addr));
+    bytes32 value = bytes32(uint256(uint160(_addr)));
     bytes memory alphabet = '0123456789abcdef';
     bytes memory str = new bytes(42);
     str[0] = '0';
